loss v45.1: cast soft cls target to cls_target dtype (AMP Half index-put)
v45 crashed at step 1 under AMP: cls_target mirrors pred_cls (Half on CUDA
autocast); float q -> dtype mismatch on index_put. Cast q at write time.

+++ losses/loss.py
@@ -256,7 +256,8 @@ class DetectionLoss(nn.Module):
                 # quality they don't have.
                 lbl = glabels[fg_gts]
                 q = align_norm[fg_gts, fg_anchors].clamp(0, 1)   # detached
-                cls_target[bi, fg_anchors, lbl] = q
+                # dtype guard: cls_target mirrors pred_cls (Half under AMP)
+                cls_target[bi, fg_anchors, lbl] = q.to(cls_target.dtype)
                 obj_target[bi, fg_anchors, 0] = 1.0
 
         # cls: balanced BCE over ALL anchors with quality-aware soft targets
